Moved app_strings into helpers.make_build
