+++ corehq/apps/app_manager/helpers/make_build/app_strings.py
